// pop-web/src/app/shared/openapi/model/genomic-signature-types.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type GenomicSignatureTypes = 'tumor_mutational_burden' | 'loss_of_heterozygosity' | 'microsatellite_instability' | 'homologous_recombination_deficiency' | 'tumor_neoantigen_burden' | 'aneuploid_score';

export const GenomicSignatureTypes = {
    TumorMutationalBurden: 'tumor_mutational_burden' as GenomicSignatureTypes,
    LossOfHeterozygosity: 'loss_of_heterozygosity' as GenomicSignatureTypes,
    MicrosatelliteInstability: 'microsatellite_instability' as GenomicSignatureTypes,
    HomologousRecombinationDeficiency: 'homologous_recombination_deficiency' as GenomicSignatureTypes,
    TumorNeoantigenBurden: 'tumor_neoantigen_burden' as GenomicSignatureTypes,
    AneuploidScore: 'aneuploid_score' as GenomicSignatureTypes
};