// pop-web/src/app/shared/openapi/model/tumor-mutational-burden-status-choices.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type TumorMutationalBurdenStatusChoices = 'low' | 'high' | 'intermediate' | 'indeterminate';

export const TumorMutationalBurdenStatusChoices = {
    Low: 'low' as TumorMutationalBurdenStatusChoices,
    High: 'high' as TumorMutationalBurdenStatusChoices,
    Intermediate: 'intermediate' as TumorMutationalBurdenStatusChoices,
    Indeterminate: 'indeterminate' as TumorMutationalBurdenStatusChoices
};