// pop-web/src/app/shared/openapi/model/tumor-mutational-burden-create.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TumorMutationalBurdenStatusChoices } from './tumor-mutational-burden-status-choices';


export interface TumorMutationalBurdenCreate { 
    category?: TumorMutationalBurdenCreate.CategoryEnum;
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Indicates the case of the patient who\'s lifestyle is assesed
     */
    caseId: string;
    /**
     * Clinically-relevant date at which the patient\'s genomic signature was assessed.
     */
    date: string;
    /**
     * The actual tumor mutational burden (TMB) value in mutations/Mb
     */
    value: number;
    status?: TumorMutationalBurdenStatusChoices | null;
}
export namespace TumorMutationalBurdenCreate {
    export type CategoryEnum = 'tumor_mutational_burden';
    export const CategoryEnum = {
        TumorMutationalBurden: 'tumor_mutational_burden' as CategoryEnum
    };
}