// pop-web/src/app/shared/openapi/model/tumor-neoantigen-burden-create.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TumorNeoantigenBurdenCreate { 
    category?: TumorNeoantigenBurdenCreate.CategoryEnum;
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Indicates the case of the patient who\'s lifestyle is assesed
     */
    caseId: string;
    /**
     * Clinically-relevant date at which the patient\'s genomic signature was assessed.
     */
    date: string;
    /**
     * The actual tumor neoantigen burden (TNB) value in neoantigens/Mb
     */
    value: number;
}
export namespace TumorNeoantigenBurdenCreate {
    export type CategoryEnum = 'tumor_neoantigen_burden';
    export const CategoryEnum = {
        TumorNeoantigenBurden: 'tumor_neoantigen_burden' as CategoryEnum
    };
}